+++ gateway/src/apicast/policy/oidc_authentication/init.lua
@@ -1,0 +1,1 @@
+return require('oidc_authentication')